requirements.txt
asgiref==3.4.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
Markdown==3.3.3
Pillow==8.4.0
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.1.1
zipp==3.6.0
